Moved modifiers into files
