+++ code/globals.lua
@@ -1,0 +1,4 @@
+m=math
+sin,cos,max,min=m.sin,m.cos,m.max,m.min
+abs,pi,rand,exp=m.abs,m.pi,m.random,m.exp
+tau=pi*2
