+++ effect/twist_fft.lua
@@ -6,7 +6,7 @@ return {
     boot=function()
     end,
     draw=function(data)
-        it=it*10*EControl
+        local it=data.et*10*EControl
         -- lets do the twist again
         for i=0,239 do
             local x=(i-it//1)%240
@@ -25,23 +25,23 @@ return {
             d=d/4
         
             if y1 < y2 then
-            line(i,cy+y1,i,cy+y2,mm(d,0,15))
+                line(i,cy+y1,i,cy+y2,clamp(d,0,15))
             end
             if y2 < y3 then
-            line(i,cy+y2,i,cy+y3,mm(d+1,0,15))
+                line(i,cy+y2,i,cy+y3,clamp(d+1,0,15))
             end
             if y3 < y4 then
-            line(i,cy+y3,i,cy+y4,mm(d+2,0,15))
+                line(i,cy+y3,i,cy+y4,clamp(d+2,0,15))
             end
             if y4 < y1 then
-            line(i,cy+y4,i,cy+y1,mm(d+3,0,15))
+                line(i,cy+y4,i,cy+y1,clamp(d+3,0,15))
             end
         end
     end,
     bdr=function(l)
         local lm=68-abs(68-l)
         for i=0,47 do
-            poke(16320+i,mm(sin(i)^2*i*lm/5.5,0,255))
+            poke(16320+i,clamp(sin(i)^2*i*lm/5.5,0,255))
         end
     end,
 }
