+++ vj80.lua
@@ -5,59 +5,30 @@
 
 package.path=package.path..";C:\\Users\\micro\\AppData\\Roaming\\com.nesbox.tic\\TIC-80\\mantratronic-vj-80\\?.lua"	-- jtruk
 
+require("code/globals")
 require("debug/fakefft")
-
-X = 7
-function X_BOOT()
-end
-
-function X_DRAW(it,ifft)
-end
-
---[[
-VolTest_DRAW(t,t)
- TwistFFT_DRAW(ET,0)
- SunBeat_DRAW(ET,0)
- FujiTwist_DRAW(ET,MID)
- ATTunnel_DRAW(ET,MID)
- Quup_DRAW(ET,MID)
- TunnelWall_DRAW(ET,MID)
- CloudTunnel_DRAW(ET,MID)
- SwirlTunnel_DRAW(ET,BASS)
- Chladni_DRAW(ET,BASS)
- CircleColumn_DRAW(ET,BASS)
- BrokenEgg_DRAW(ET,BASS)
- ParaFlower_DRAW(ET,BASS)
- FFTCirc_DRAW(ET,BASS)
- Proxima_DRAW(ET,BASS)
- Lemons_DRAW(ET,BASS)
- RevisionBack_DRAW(ET,BASS)
- Bitnick_DRAW(ET,BASS)
- Worms_DRAW(ET,BASS)
- --]]
+FontBoot=require("code/font")
 
 -- effects follow the template: {boot=function(), draw=function(), bdr=function()}
 Effects={
   require("effect/vol_test"),
   require("effect/twist_fft"),
-  --[[
-		require("effect/sun_beat"),
-		require("effect/fuji_twist"),
-		require("effect/at_tunnel"),
-		require("effect/quup"),
-		require("effect/tunnel_wall"),
-		require("effect/cloud_tunnel"),
-		require("effect/swirl_tunnel"),
-		require("effect/chladni"),
-		require("effect/circle_column"),
-		require("effect/broken_egg"),
-		require("effect/para_flower"),
-		require("effect/fft_circ"),
---]]
+  require("effect/sun_beat"),
+--  require("effect/fuji_twist"),
+  require("effect/at_tunnel"),
+  require("effect/quup"),
+  require("effect/tunnel_wall"),
+  require("effect/cloud_tunnel"),
+  require("effect/swirl_tunnel"),
+  require("effect/chladni"),
+  require("effect/circle_column"),
+  require("effect/broken_egg"),
+  require("effect/para_flower"),
+  require("effect/fft_circ"),
   require("effect/proxima"),
   require("effect/lemons"),
   require("effect/revision_back"),
-----		require("effect/bitnick"),
+  require("effect/bitnick"),
   require("effect/worms"),
 }
 
@@ -99,24 +70,32 @@ Overlays = {
 --]]
 }
 
---NumEffects=16
---NumOverlays=14
---NumModifiers=11
+Modifiers={
+  require("modifier/pix_noise"),
+  require("modifier/pix_zoom"),
+  require("modifier/grid_dim"),
+  require("modifier/post_circ"),
+  require("modifier/pix_motion_blur"),
+  require("modifier/pix_jump_blur"),
+  require("modifier/rot_vert"),
+  require("modifier/rot_horz"),
+  require("modifier/sft_vert"),
+  require("modifier/sft_horz"),
+  require("modifier/post_squares"),
+  require("modifier/line_scratch"),
+}
+
 NumModes=8
 NumPalettes=13
 
 -- Utils
 
-m=math
-sin=m.sin cos=m.cos max=m.max min=m.min
-abs=m.abs pi=m.pi tau=m.pi*2 rand=m.random
-
-function mm(x,a,b)
+function clamp(x,a,b)
  return max(a,min(b,x))
 end
 
 function ss(x,e1,e2)
- y=mm(x,e1,e2)
+ y=clamp(x,e1,e2)
  st=(y-e1)/(e2-e1)
  return st*st*(3-2*st)
 end
@@ -148,42 +127,6 @@ function printlogo(x,y,kx,ky,col)
    print(string.sub(logo[i],ch,ch),x+(ch-1)*kx,y+(i-1)*ky,col+i,true,1,true)
   end
  end
-end
-
-local rle = "0800020ODODOHPDPHPHPHPHP4HPPHOHOHAHOHOHOHAHOHOHOHAHOP2HHAHOP2DHAHOPHOPOHHOPHP2HHOPHP2HHOHAHAHAHOHAHAHAHOPDPDHP3DPDHP3DPDHP2ODOPHOHAODOPHOHAODOPHOHAMBAOHOHAMBAOHOHAMBAOPHHAMBAOPDHAMBAOPHHAMPDAPDMHOPDAPHOP2DAP3HLDAHOHOHLDAHOHOHLDAHOHOHLDAHOHOHIDAHOHOP2HHAHOHOHOHAHOHOHOHAHOHOHOHAHOHOP5HOPHOP2HOPDMPPHA7HAHAHOHOHAHAHOHOHAHAHOHOHAHAHOHOPPHAPPHOPPHAOPHOOPHAMPHOA7MBAOHOHAMBAOHOHAMBAOHOHAMBAOHOHAPHPPHOP2HPPHOP2HPHHOOPA7HIDAHOHOHIDAHOHOHIDAHOHOHIDAHOHOHIDAHOPPHIDAHOPHHIDAHOODA7ODMHODMPPHOPPHOP8HOHOHOHAHOHOHOHAHOHOHOPDPPHOPHOHPHHOPDMP2BAHOHOPPBAHOHOPPBAHOHOIDAAHOHOIDAAHOHOIDAAHOHOIDAAHOHOIDAAHOHOHIDAHOH2IDAHOH2IDAHOH2IDAHOH2IDAHOH2IDAOH3IDAMDPHHIDAOHPHPPAAHA2PPAAHA2PPAAHA3OAAHA3OPDPDOHMPPHPHPHOHP4HPDAOHOHAPDHOPHAOHAHOHOAOHAHOHOAOHAHPHOAOHAPHHOPPHAPPHOPPHAONHOPHA7IDAAHOHOIDAAHOHOIDAAHOHOIDAAHOHOIDAAP3IDAAOPPHIDAAMPPDA7HLDAHOODHLDAHOMBHLDAHOMBHLDAHOMBPPDAHOMBOPDAHOMBMPBAHOMBA7HAOPHOHAHAPPHOHAHAPPHOHAHAHOHOHAP13OP2OPPHMPA8OAAMPA2OAAOPA2OAAOPA2OAAOA2MPOHPDMPOP3DOP5DPPHOHOOAHOHA4HAHAMBMBHAHAMBMBHAHA4HAPDPAPAHOPHPBPBHOP2BPBHOHOMBMBHOHA6HA6HA6HA6HAMPBAPDHAOPDAPHHAPPDAPPHAHLDAHOA31MHODMHMHOPPHOPOHP6HHOHOHOHAHOPPOAHOHOPPOAHOHOPHOAHOHOHAOAHOP3OAP3OPOAOPOPMPOAMPA6OHOMBMBPPHOMBMBPHHOMBMBPPHOMBMBHOHOMBMBHOHOMBMBHOHOMBMBHOA3MBAAHAHLDAHOHAHLDAHOHAHIDAHOHAHIDAHOPDHIDAHOPDHIDAHOODHIDAHOA7HOHOHOHAHOHOHOHAHOHOHOHAHOHOHOHAP5HAPHPHOPHAODPDMPHA2HAAOA7OPA5OHA5ODA43PBA5PBA5PA108HAAOA3HAAOA3HAAOA43HA6HA6HA6HA4MPPDHOHOOPPDHOHOP2DHOHOHAHAHOHOA31HIDAHOGOHIDAHOHOHIDAHOHOHIDAHOHOAAGAHA4HAHA4HAHA4HADA2PPHA4PPHA2MBPPHA2MBAOHA2PHA5IDA5IDA5IDA5MDA5MBA5MBA5MBAAPHAAOBPDHAHOHOOHHAHOHOMPHAHOHOAOHAHOHOP2HP5OHOPPHPHMHMPPDA7HIDAPPHOHLDAOHHOHLDAPPHOHLDAHOHOPPDAHOPPOPDAHOOPMPBAHOMPA6OMPHA2PHOHHA2PHPDHA2MBHA4MBPPHA4PPHA4PPHA14PHAAOA2PHAAOA6OA6PAHA2HAHAHA2HAHAHA2HAHADA76OPA5OHA5ODA167MHPBPHPHOPPDP6DP3HOIDAOAOHPIDAOAOHPIDMPMP2IDOHMHPPIDPDMPHOPPMP2HOPPOP2HOP5HOHAHAAOHOHAHAAOHOPDPDAP3HPHIHOPOP2MDMHMHA3OPOPA3P3A3HOHOA3HOHOA3PPHOHAPHOHPPHAPHP3HAPHPHIJDA2PPMJDA2PPMJDA3OP2A2MP3A2OHP2A2PDMJDA2HAMJDA2POIDHAAOPOIDHAAOHOIDHAAOHOIDHAAOPPIDP4HIDP2HODIDP2DA7MPAOHOMBAOAOHOMBAOAOHOMBAOAOHOMBAOP3MBAOPPOHMBAOPHMDMBA7HOOPHA2HOAOAAPHHOAOAAPHHOAOHAPHP3HA2PHPHHA2ODPDHA10PPMJDA2P4A2OP3A4P2A2MBMJDA2MBMJDA2MBMJBA10"
--- font data {"A", sprite number, page?, num sprites x, y, width (px), height}
-fontd={	{0,0,1,2,8,16},{1,0,1,2,8,16},{2,0,1,2,8,16},{3,0,1,2,8,16},{4,0,1,2,8,16},{5,0,1,2,8,16},{6,0,1,2,8,16},{7,0,1,2,8,16},{8,0,1,2,7,16},{9,0,1,2,8,16},{10,0,1,2,8,16},{11,0,1,2,8,16},{12,0,2,2,10,16},{14,0,1,2,8,16},{15,0,1,2,8,16},
-		{128,0,1,2,8,16},{129,0,1,2,8,16},{130,0,1,2,8,16},{131,0,1,2,8,16},{132,0,2,2,9,16},{134,0,1,2,8,16},{135,0,1,2,8,16},{136,0,2,2,10,16},{138,0,1,2,8,16},{139,0,1,2,7,16},{140,0,1,2,8,16},{141,0,1,2,8,16},{142,0,1,2,8,16},{143,0,1,2,8,16},
-		{256,0,1,2,8,16},{257,0,1,2,8,16},{258,0,1,2,8,16},{259,0,1,3,8,19},{260,0,1,2,8,16},{261,0,1,2,5,16},{262,0,1,3,6,19},{263,0,1,2,8,16},{264,0,1,2,6,16},{265,0,2,2,10,16},{267,0,1,2,8,16},{268,0,1,2,8,16},{269,0,1,3,8,19},{270,0,1,3,8,19},{271,0,1,2,7,16},
-		{448,0,1,2,8,16},{449,0,1,2,7,16},{450,0,1,2,8,16},{451,0,1,2,8,16},{452,0,2,2,10,16},{454,0,1,2,8,16},{455,0,1,3,8,19},{456,0,1,2,8,16},{457,0,1,2,3,16},{458,0,1,2,3,16},{459,0,1,2,7,16},{460,0,1,2,3,16},{461,0,1,2,7,16},{462,0,1,2,3,16},{463,0,1,2,6,16},
-		{640,0,1,2,8,16},{641,0,1,2,6,16},{642,0,1,2,8,16},{643,0,1,2,8,16},{644,0,1,2,8,16},{645,0,1,2,8,16},{646,0,1,2,8,16},{647,0,1,2,8,16},{648,0,1,2,8,16},{649,0,1,2,8,16},{650,0,1,2,3,16},{651,0,1,2,7,16},{652,0,1,2,8,16},{653,0,2,2,12,16},{655,0,1,2,8,16}
-	}
-
--- this could be useful for compression later
-font = {}
-chars="ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz!\'+,-./0123456789:=?# "
-
--- the rle decoder
-function tomemrle(str)
-  local o=tonumber(str:sub(1,5),16) -- get (o)ffset
-  local w=tonumber(str:sub(6,7),16)*8-1 -- get (w)idth
-  local e=str:sub(8,str:len()) -- remove header to get (e)ncoded data
-  local d = "" -- (d)ecoded data
-  for m, c in e:gmatch("(%u+)([^%u]+)") do -- decode rle, (m)atch & (c)ounter
-    d = d .. m .. (m:sub(-1):rep(c))  
-  end
-  local y=0
-  for x = 1,#d,1 do -- write to mem
-    local c=string.byte(d:sub(x,x))-65 -- get (c)olor value
-    poke4(o+y,c) y=y+1
-    if y>w then y=0 o=o+1024 end
-  end
-end
-
-function font_init()
-	for i=1,#fontd do
-		font[string.sub(chars,i,i)] = fontd[i]
-	end
 end
 
 function flength(txt,kx,size)
@@ -222,11 +165,6 @@ function fprint(txt,tx,ty,kx,ky,tc,size)
 	end
 end
 
-function FONT_BOOT()
-    -- initialize
-    tomemrle(rle)
-    font_init()
-end
 
 
 -- FFT setup
@@ -280,547 +218,30 @@ function FFT_FILL()
  --high=high/4
 end
 
-OL_lines = {}
-OL_drawlines = {}
-OL_frames = {}
-OL_width=16
-OL_height=120
-OL_numframes=240
-
-function OL_scan(ilines)
-  OL_drawlines={}
-  OL_frames={}
-  for j=1,OL_numframes do
-    OL_drawlines={}
-   a=j/OL_numframes*tau
-   for i=1,#ilines do
-    ln=ilines[i]
-    cx=ln.cx
-    cy=ln.cy
-    x1=cx-ln.r/2
-    x2=cx+ln.r/2
-    z1=-OL_width/2
-    z2=OL_width/2
-    
-    a1=sin(a-tau/8)
-    a2=sin(a+tau/8)
-    a3=sin(a+tau*3/8)
-    a4=sin(a-tau*3/8)
- 
-    X1=x1*cos(a)-z2*sin(a)
-    Z1=x1*sin(a)+z2*cos(a)
-    X2=x1*cos(a)-z1*sin(a)
-    Z2=x1*sin(a)+z1*cos(a)
-    X3=x2*cos(a)-z1*sin(a)
-    Z3=x2*sin(a)+z1*cos(a)
-    X4=x2*cos(a)-z2*sin(a)
-    Z4=x2*sin(a)+z2*cos(a)
-    if(X1 < X2) then
-     c=(1-(abs(a%tau-tau/4)/tau))*15
-     table.insert(OL_drawlines,{X1,cy,X2,cy,c,(Z1+Z2)/2})
-    end
-    if(X2 < X3) then
-     c=(1-(abs(a%tau-tau/2)/tau))*15
-     table.insert(OL_drawlines,{X2,cy,X3,cy,c,(Z2+Z3)/2})
-    end
-    if(X3 < X4) then
-     c=(1-(abs(a%tau-tau*3/4)/tau))*15
-     table.insert(OL_drawlines,{X3,cy,X4,cy,c,(Z3+Z4)/2})
-    end
-    if(X4 < X1) then
-     c=(1-((abs(a%tau-tau/2))/tau))*15
-     table.insert(OL_drawlines,{X4,cy,X1,cy,c,(Z4+Z1)/2})
-    end
-   end
-   table.sort(OL_drawlines, function (a,b) return a[6] < b[6] end)
-   OL_frames[j]=OL_drawlines
-  end
-
-  table.insert(OldLogos,OL_frames)
-end
-
-function vectrex_BOOT()
-  vec_lines={}
-  cls()
-  for y=10,Fuji_height-10 do
-    line(230-y,y,10+y,y,1)
-  end
-  for y=10,Fuji_height-40 do
-    line(30+y,y,30+Fuji_width+y,y,0)
-  end
-  for y=1,Fuji_height do
-    -- fuck it search by image
-    c=1
-    for x=0,240 do
-     if c==1 and pix(x,y) == 1 then
-      c1=x
-      c=2
-     elseif c==2 and pix(x,y) == 0 then
-      c2=x
-       
-      r=c1-c2
-      l={cx=r/2+x-120, cy=y+10, r=r}
-      table.insert(vec_lines,l)
-      c=1
-     end
-    end
-  end
-  OL_scan(vec_lines)
-end
-
-function amiga_BOOT()
-  amiga_lines={}
-  cls()
-  for y=10,Fuji_height-10 do
-    line(200-y,y,200+Fuji_width-y,y,1)
-
-    line(177-y,y,177+Fuji_width-y,y,1)
-  end
-
-  i=0
-  for y=Fuji_height-10,Fuji_height-50,-1 do
-    i=i+2
-    line(177-y-i,y,177-y-i+Fuji_width,y,1)
-
-    line(200-y-i,y,200-y-i+Fuji_width,y,1)
-  end
-  for y=1,Fuji_height do
-    -- fuck it search by image
-    --space=136-Fuji_height
-    c=1
-    for x=0,240 do
-     if c==1 and pix(x,y) == 1 then
-      c1=x
-      c=2
-     elseif c==2 and pix(x,y) == 0 then
-      c2=x
-       
-      r=c1-c2
-      l={cx=r/2+x-120, cy=y+10, r=r}
-      table.insert(amiga_lines,l)
-      c=1
-     end
-    end
-  end
-    OL_scan(amiga_lines)
-
-end
-
-function c64_BOOT()
-  c64_lines={}
-  c64_drawlines={}
-  c64_frames={}
-  l1y=Fuji_height*4/12
-  l2y=Fuji_height*4/8+Fuji_height/16
-  
-  numpoints=Fuji_height*10
-  p1={}
-  p2={}
-  
-  cls()
-  circ(110,52,50,1)
-  circ(110,52,34,0)
-  rect(120,0,120,136,0)
-  
-  rect(120,36,34,14,1)
-  rect(120,54,34,14,1)
-  
-  tri(140,52,160,32,160,72,0)
-  
-  for y=1,Fuji_height do
-   -- fuck it search by image
-   space=136-Fuji_height
-   c=1
-   for x=0,240 do
-    if c==1 and pix(x,y) == 1 then
-     c1=x
-     c=2
-    elseif c==2 and pix(x,y) == 0 then
-     c2=x
-      
-     r=c1-c2
-     l={cx=r/2+x-120, cy=y+space, r=r}
-     table.insert(c64_lines,l)
-     c=1
-    end
-   end
-  end
-
-  OL_scan(c64_lines)
- end
---]]
-
-
-
 -- Modifiers
 
 OOrder = 0
 EOrder = 0
 function ModifierHandler(COrder,IOrder,mod, MT,MC)
-		local Mod = Modifiers[mod]
+	local modifier = Modifiers[mod]
   if COrder == IOrder then
-    if Mod == 0 then
-    elseif Mod == PIXNoise then
-     PIXNoise_DRAW(1000*HIGH,MC)
-    elseif Mod == PIXZoom then
-     PIXZoom_DRAW(1000*HIGH,MC)
-    elseif Mod == GridDim then
-     GridDim_DRAW(1000*HIGH, MT,MC)
-    elseif Mod == POSTCirc then
-     POSTCirc_DRAW(1000*HIGH, MT,MC)
-    elseif Mod == PIXMotionBlur then
-      PIXMotionBlur_DRAW(1000*HIGH,MT,MC)
-    elseif Mod == PIXJumpBlur then
-      PIXJumpBlur_DRAW(1000*HIGH,MT,MC)
-    elseif Mod == ROTVert then
-      ROTVert_DRAW(1000*HIGH,MT,MC)
-    elseif Mod == ROTHorz then
-      ROTHorz_DRAW(1000*HIGH,MT,MC)
-    elseif Mod == SFTVert then
-      SFTVert_DRAW(1000*HIGH,MT,MC)
-    elseif Mod == SFTHorz then
-      SFTHorz_DRAW(1000*HIGH,MT,MC)
-    elseif Mod == POSTSquares then
-      POSTSquares_DRAW(1000*HIGH,MT,MC)
-    elseif Mod == EvilpaulGlitch then
-      EvilpaulGlitch_DRAW(1000*HIGH,MT,MC)
-				elseif Mod == LineScratch then
-      LineScratch_DRAW(1000*HIGH,MT,MC)
+    if modifier ~= nil then
+      modifier.draw({amount=1000*HIGH, mt=MT, mc=MC})
     end
   end
-end
-
-PIXNoise = 1
-function PIXNoise_DRAW(amount, mc)
- for i=0,amount do
-  x=rand(240)-1
-  y=rand(136)-1
-  pix(x,y,mm(pix(x,y)-mc,0,15))
- end
-end
-
-PIXZoom = 2
-
-function PIXZoom_DRAW(amount, mc)
- d=1+2*rand()
- for i=1,amount do
-  x=240*rand()
-  y=136*rand()
-  a=math.atan(x-120,y-68)
-
-  op=pix(x,y)-mc
-  if op >= 0 then
-   pix(x+d*(sin(a)+sin(t/300)),y+d*(cos(a)+sin(t/300)),op)
-  else
-   pix(x+d*sin(a),y+d*cos(a),0)
-  end
- end
-end
-
-GridDim = 3
-
-function GridDim_DRAW(amount, mt, mc)
- i=0
- for y=-1,36 do
-  for x=-1,60 do
-   i=i+1
-   --if i > amount*5 then return end
-   sx=x*4+(mt*20)%4
-   sy=y*4+(mt*5)%4
-   pix(sx,sy,mm(pix(sx,sy)-mc,0,15))
-  end
- end
-end
-
-POSTCirc = 4
-
-function POSTCirc_DRAW(amount, mt,mc)
-  it=mt*40
-  lim = mm(8+mc,0,15)
-  for y=0,135 do 
-   for x=0,239 do
-    Y=(68-y)
-    X=(120-x)
-    d=math.sqrt(X^2+Y^2)//1
-    a=math.atan2(X,Y)
-    if((10*sin(it/4+d/5))%100>50) then
-     c=pix(x,y)
-     if c>lim then
-      pix(x,y,mm(c-8,0,15))
-     end
-    end 
-   end 
-  end
-end
-
-PIXMotionBlur = 5
-PMBsize = 20
-function PIXMotionBlur_DRAW(amount, mt, mc)
- size=PMBsize+(mt)%5
- limit=50 + mc
- for i=0,amount/4 do
-  d=2+size+rand(limit)
-  a=rand()*tau
-  x=d*sin(a)
-  y=d*cos(a)
-  if x >= -119 and x <= 118 and y >=-67 and y <= 66 then
-   pix(120+x,68+y,mm(((pix(120+x,68+y)+pix(120+x-1,68+y-1)+pix(120+x+1,68+y-1)+pix(120+x-1,68+y+1)+pix(120+x+1,68+y+1))/4.8),0,15))
-  end
- end
-    
- for i=0,amount do
-  d=size+rand(limit)
-  a=rand()*tau
-  x=d*sin(a)
-  y=d*cos(a)
-  if x >= -120 and x <= 119 and y >=-68 and y <= 67 then
-   pix(120+x,68+y,pix(120+(d-1)*sin(a),68+(d-1)*cos(a)))
-  end
- end
-end
-
-PIXJumpBlur = 6
-function PIXJumpBlur_DRAW(amount, mt, mc)
-  size=10+(mt)%5
-  limit=100 + mc
-  cx=120 cy=68
-  tt=BASS%1
-  for i=0,tt*50 do
-    d=rand()
-    d=1-(d^1.5)
-    d1=size+d*(limit-tt*10)
-    d2=1+d1+tt*10
-    a=rand()*tau
-    x=d1*sin(a)
-    y=d1*cos(a)
-    line(cx+(d1)*sin(a),cy+(d1)*cos(a),cx+(d2)*sin(a),cy+(d2)*cos(a),pix(cx+x,cy+y))
-  end
-     
-  amount = min(amount,500)
-  for i=0,amount do
-   d=rand()
-   d=1-(d^1.5)
-   d=d*(size//1+limit*1.5)
-   a=rand()*tau
-   x=cx+d*sin(a)
-   y=cy+d*cos(a)
-   if x >= 1 and x <= 239 and y >=1 and y <= 134 then
-    pix(cx+(d+1)*sin(a),cy+(d+1)*cos(a),mm((pix(x,y)+pix(x+1,y+1)+pix(x+1,y-1)+pix(x-1,y+1)+pix(x-1,y-1))/4.8,0,15))
-   end
-  end
-end
- 
-ROTVert = 7
-function ROTVert_DRAW(amount,mt,mc)
-  
-  dir=1
-  local lines = 0
-  if mc == 0 then
-    lines = mt%5//1
-  else
-    if mc < 0  then
-      dir = -1
-    end
-    --lines = abs(mc)*(mt%4+1)//1
-    lines = (abs(mc)*(mt+1)//1)%136
-  end
-
-  if dir == 1 then
-    -- going down
-    memcpy(0x8000,(135-lines)*120,120*lines)
-    for y=135-lines,0,-1 do
-     memcpy((y+lines)*120,y*120,120)
-    end
-    memcpy(0,0x8000,120*lines)
-  elseif dir == -1 then
-    -- going up
-    memcpy(0x8000,0,120*lines)
-    for y=0,135-lines do
-      memcpy(y*120,(y+lines)*120,120)
-     end
-     memcpy((136-lines)*120,0x8000,120*lines)
- 
-  end
-end
-
-ROTHorz = 8
-function ROTHorz_DRAW(amount,mt,mc)
-  dir=1
-  local pixels = 0
-  if mc == 0 then
-    pixels = mt%5//1
-  else
-    if mc < 0  then
-      dir = -1
-    end
-    --pixels = abs((mc*mt)%120)
-    --pixels = (abs(mc)*mt)%120
-    pixels = (abs(mc)*(mt+1)//1)%120
-  end
-
-  if dir ==1 then
-    -- going right
-    for y=0,135 do
-      -- take the whole line
-      memcpy(0x8000,y*120, 120)
-
-      -- put it back in two sections
-      memcpy(y*120+pixels,0x8000, 120-pixels)
-      memcpy(y*120,0x8000+(120-pixels), pixels)
-    end
-  else
-    -- going left
-    for y=0,135 do
-      -- take the whole line
-      memcpy(0x8000,y*120, 120)
-
-      -- put it back in two sections
-      memcpy(y*120+(120-pixels),0x8000, pixels)
-      memcpy(y*120,0x8000+pixels, 120-pixels)
-    end
-  end
-
-end
-
-SFTVert = 9
-function SFTVert_DRAW(amount,mt,mc)
-  dir=1
-  local lines = 0
-  if mc == 0 then
-    lines = mt%5//1
-  else
-    if mc < 0  then
-      dir = -1
-    end
- --    lines = abs(mc)*(mt%4+1)//1
-    lines = (abs(mc)*(mt+1)//1)%136
-  end
-
-  if dir == 1 then
-    -- going down
-    for y=135-lines,0,-1 do
-     memcpy((y+lines)*120,y*120,120)
-    end
-    memset(0,0,120*lines)
-  elseif dir == -1 then
-    -- going up
-    for y=0,135-lines do
-      memcpy(y*120,(y+lines)*120,120)
-     end
-     memset((136-lines)*120,0,120*lines)
-  end
-end
-
-SFTHorz = 10
-function SFTHorz_DRAW(amount,mt,mc)
-  dir=1
-  local pixels = 0
-  if mc == 0 then
-    pixels = mt%5//1
-  else
-    if mc < 0  then
-      dir = -1
-    end
-    --pixels = abs((mc*mt)%120)
-    --pixels = (abs(mc)+(mt*2))%120
-    pixels = (abs(mc)*(mt+1)//1)%120
-  end
-
-  if dir ==1 then
-    -- going right
-    for y=0,135 do
-      -- take the whole line
-      memcpy(0x8000,y*120, 120)
-
-      -- put it back in two sections
-      memcpy(y*120+pixels,0x8000, 120-pixels)
-      memset(y*120,0, pixels)
-    end
-  else
-    -- going left
-    for y=0,135 do
-      -- take the whole line
-      memcpy(0x8000,y*120, 120)
-
-      -- put it back in two sections
-      memset(y*120+(120-pixels),0, pixels)
-      memcpy(y*120,0x8000+pixels, 120-pixels)
-    end
-  end
-
-end
-
-POSTSquares = 11
-PSp={x=0,y=0,sx=10,sy=10,t=0.02,lt=0}
-function POSTSquares_DRAW(amount,mt,mc)
- if abs(mt-PSp.lt) >= PSp.t then
-  PSp.lt=mt
-  grid=240/PSp.sx
-  PSp.x=(rand(grid//1)-1)*PSp.sx
-  grid=136/PSp.sy
-  PSp.y=(rand(grid//1)-1)*PSp.sy
- end
- for i=PSp.x,PSp.x+PSp.sx do
-  for j=PSp.y,PSp.y+PSp.sy do
-    pix(i,j,mm(pix(i,j)-mc,0,15))
-  end
- end
-end
-
-EvilpaulGlitch = 12
-function EvilpaulGlitch_DRAW(amount,mt,mc)
-	 math.randomseed(t//60)
-  for i = 0, math.random(0, 2) do
-     local x1 = math.random(0, 240)
-     local x2 = math.random(0, 240)
-     local w = math.random(5, 30)
-     local m = math.random(0, 1)
-     if m == 0 then
-      for x = 0, w do
-       for y = 0, 136 do
-        local c = pix(x1 + x, y)
-        pix(x2 + x, y, c)
-       end
-      end
-     elseif m == 1 then
-      for y = 0, 136 do
-       local c = pix(x1, y)
-       for x = 0, w do
-        pix(x2 + x, y, c)
-       end
-      end
-     elseif m == 2 then
-      for y = 0, 136 do
-       for x = 0, w do
-     		 pix(x2 + x, y, math.random(0, 1))
-       end
-      end
-     end
-		 end	
-end
-
-LineScratch = 13
-function LineScratch_DRAW(amount,mt,mc)
-	for a=0,46 do
-		local x=math.random(240)
-		local y=math.random(136)
-		local w=math.random(20)
-		line(x,y,x+w,y,pix(x,y))
-		line(x,y+1,x+w/2,y+1,pix(x,y))
-	end
 end
 
 Texts={
-						{"LOVEBYTE","2024","+ + +","GOTO80","+ + +"},
-      {"ACID","DANCE","ROBOT","NINJA"},
-      {"LOVE","LIFE","LIVE","LEFT","LOUD","LINE"},
-      {"MATH","SINE","CIRC","LINE","CLS","RECT"},
-      {"MUNCH","---"},
-      {"I","DONT","KNOW"},
-      {"SHALALA","---"},
-      {"HARDCORE","FAMILY"},
-      {"BREAK","BEAT","DANCE","HIT","HOP","SHOUT","SCREAM","JUMP"}
-      }
+  {"LOVEBYTE","2024","+ + +","GOTO80","+ + +"},
+  {"ACID","DANCE","ROBOT","NINJA"},
+  {"LOVE","LIFE","LIVE","LEFT","LOUD","LINE"},
+  {"MATH","SINE","CIRC","LINE","CLS","RECT"},
+  {"MUNCH","---"},
+  {"I","DONT","KNOW"},
+  {"SHALALA","---"},
+  {"HARDCORE","FAMILY"},
+  {"BREAK","BEAT","DANCE","HIT","HOP","SHOUT","SCREAM","JUMP"}
+}
 TImages={}
 
 -- TestSheet
@@ -849,9 +270,9 @@ PAL_CURRENT=0
 function PAL_make2(r1,g1,b1,r2,g2,b2)
  local pal={}
  for i=0,15 do
-  pal[i*3]   = mm(r1*abs(15-i)/15 + r2*abs(i)/15,0,255)
-  pal[i*3+1] = mm(g1*abs(15-i)/15 + g2*abs(i)/15,0,255)
-  pal[i*3+2] = mm(b1*abs(15-i)/15 + b2*abs(i)/15,0,255)
+  pal[i*3]   = clamp(r1*abs(15-i)/15 + r2*abs(i)/15,0,255)
+  pal[i*3+1] = clamp(g1*abs(15-i)/15 + g2*abs(i)/15,0,255)
+  pal[i*3+2] = clamp(b1*abs(15-i)/15 + b2*abs(i)/15,0,255)
  end
  return pal
 end
@@ -859,14 +280,14 @@ end
 function PAL_make3(r1,g1,b1,r2,g2,b2,r3,g3,b3)
  local pal={}
  for i=0,7 do
-  pal[i*3]   = mm(r1*abs(7-i)/7 + r2*abs(i)/7,0,255)
-  pal[i*3+1] = mm(g1*abs(7-i)/7 + g2*abs(i)/7,0,255)
-  pal[i*3+2] = mm(b1*abs(7-i)/7 + b2*abs(i)/7,0,255)
+  pal[i*3]   = clamp(r1*abs(7-i)/7 + r2*abs(i)/7,0,255)
+  pal[i*3+1] = clamp(g1*abs(7-i)/7 + g2*abs(i)/7,0,255)
+  pal[i*3+2] = clamp(b1*abs(7-i)/7 + b2*abs(i)/7,0,255)
  end
  for i=1,8 do
-  pal[21+i*3]   = mm(r2*abs(8-i)/8 + r3*abs(i)/8,0,255)
-  pal[21+i*3+1] = mm(g2*abs(8-i)/8 + g3*abs(i)/8,0,255)
-  pal[21+i*3+2] = mm(b2*abs(8-i)/8 + b3*abs(i)/8,0,255)
+  pal[21+i*3]   = clamp(r2*abs(8-i)/8 + r3*abs(i)/8,0,255)
+  pal[21+i*3+1] = clamp(g2*abs(8-i)/8 + g3*abs(i)/8,0,255)
+  pal[21+i*3+2] = clamp(b2*abs(8-i)/8 + b3*abs(i)/8,0,255)
  end
  return pal
 end
@@ -899,7 +320,7 @@ function PAL_BOOT()
   for i=0,47 do
     Sweetie16PAL[i]=peek(0x3fc0+i)
 
-    BlueOrangePAL[i]=mm(sin(i)^2*i,0,255)
+    BlueOrangePAL[i]=clamp(sin(i)^2*i,0,255)
 
     BlueGreySinePAL[i]=sin(i/15)*255
 
@@ -907,9 +328,9 @@ function PAL_BOOT()
   end
 
   for i=0,15 do
-    ReddishPAL[i*3]=mm(i*32,0,255)
-    ReddishPAL[i*3+1]=mm(i*24-128,0,255)
-    ReddishPAL[i*3+2]=mm(i*24-256,0,255)
+    ReddishPAL[i*3]=clamp(i*32,0,255)
+    ReddishPAL[i*3+1]=clamp(i*24-128,0,255)
+    ReddishPAL[i*3+2]=clamp(i*24-256,0,255)
 
     OverBrownPAL[i*3]=math.min(255,20+i*32)
     OverBrownPAL[i*3+1]=math.min(255,10+i*24)
@@ -954,7 +375,7 @@ function PAL_Switch(ip, speed, buffer)
   PAL_time = 0
  end
 
- PAL_time = mm(PAL_time + speed,0,1)
+ PAL_time = clamp(PAL_time + speed,0,1)
 
  for i=0,47 do
   --ic=peek(0x3fc0+i)
@@ -965,7 +386,7 @@ function PAL_Switch(ip, speed, buffer)
   end
   nc = ip[i] 
 
-  poke(0x3fc0+i, mm(ic * (1-PAL_time) + nc *PAL_time,0,255))
+  poke(0x3fc0+i, clamp(ic * (1-PAL_time) + nc *PAL_time,0,255))
  end
 
  --sprint("pt:"..PAL_time.."|old:"..PAL_olde.."|new:"..PAL_currente.."|buffer:"..buffer.."|s:"..speed,10,10,15)
@@ -977,7 +398,7 @@ BlueOrangePAL={}
 function PAL_Fade(ip,l)
  local lm=68-abs(68-l)
  for i=0,47 do
-  poke(0x3fc0+i, mm(ip[i]*lm/5.5,0,255))
+  poke(0x3fc0+i, clamp(ip[i]*lm/5.5,0,255))
  end
 end
 
@@ -998,9 +419,9 @@ function PAL_Rotate2(it,l)
   gradeg=sin(it*1/13+l/150)+1
   gradeb=sin(it*1/11+l/150)+1
   for i=0,15 do
-   poke(0x3fc0+i*3,  mm(i*16*(grader),0,255))
-   poke(0x3fc0+i*3+1,mm(i*16*(gradeg),0,255))
-   poke(0x3fc0+i*3+2,mm(i*16*(gradeb),0,255))
+   poke(0x3fc0+i*3,  clamp(i*16*(grader),0,255))
+   poke(0x3fc0+i*3+1,clamp(i*16*(gradeg),0,255))
+   poke(0x3fc0+i*3+2,clamp(i*16*(gradeb),0,255))
   end
 end
 
@@ -1008,11 +429,11 @@ Dimmed = 7
 function PAL_Rotate3(it,l)
  for i=0,15 do
   r=i*(8+8*(sin(tau/6*5+it+l/100)))
-  poke(0x3fc0+i*3,mm(r,0,255))
+  poke(0x3fc0+i*3,clamp(r,0,255))
   g=i*(8+8*(math.sin(it+l/100)))
-  poke(0x3fc0+i*3+1,mm(g,0,255))
+  poke(0x3fc0+i*3+1,clamp(g,0,255))
   b=i*(8+8*(math.cos(it+l/100)))
-  poke(0x3fc0+i*3+2,mm(b,0,255))
+  poke(0x3fc0+i*3+2,clamp(b,0,255))
  end
 end    
 
@@ -1501,7 +922,7 @@ function KEY_CHECK()
    OL_ID = OL_ID - 1
   end
   
-  OL_ID = mm(OL_ID%(#OldLogos+1),1,#OldLogos)
+  OL_ID = clamp(OL_ID%(#OldLogos+1),1,#OldLogos)
 
   -- home: FFTH_length up by 1
   if keyp(56) then
@@ -1512,7 +933,7 @@ function KEY_CHECK()
  if keyp(57) then
   FFTH_length = FFTH_length - 1
  end 
- FFTH_length = mm(FFTH_length,2,60)
+ FFTH_length = clamp(FFTH_length,2,60)
 
  -- insert: stutter effect cls switch
  if keyp(53) == true then
@@ -1537,7 +958,7 @@ function KEY_CHECK()
 
  -- Beat detection/input
  if keyp(48,10000,10) == true then
-  local currentbeat=mm((BTC+1)%BEATS,1,BEATS)
+  local currentbeat=clamp((BTC+1)%BEATS,1,BEATS)
   BTA[currentbeat]= time()-LBT
   LBT=time()
   BTC=currentbeat
@@ -1558,7 +979,7 @@ function KEY_CHECK()
    Loudness = Loudness - 0.1
  end
 
- Loudness = mm(Loudness,0.1,10)
+ Loudness = clamp(Loudness,0.1,10)
 
  -- up: increase text image
  if keyp(58) then
@@ -1570,7 +991,7 @@ function KEY_CHECK()
   TIimageID = TIimageID + 1
  end
 
- TIimageID = mm(TIimageID%(#TImages+1),1,#TImages)
+ TIimageID = clamp(TIimageID%(#TImages+1),1,#TImages)
 
  -- left: decrease text image
  if keyp(60) then
@@ -1584,7 +1005,7 @@ function KEY_CHECK()
   if TextID > #Texts then TextID = 1 end
  end
   
- --TextID = mm(TextID%(#Texts+1),1,#Texts)
+ --TextID = clamp(TextID%(#Texts+1),1,#Texts)
   
  -- insert: effect cls switch
  if keyp(53) == true then
@@ -1616,7 +1037,7 @@ function KEY_CHECK()
   if Effect > #Effects then Effect = 1 end
  end
  
- --Effect = mm(Effect%(#Effects+1),0,#Effects)
+ --Effect = clamp(Effect%(#Effects+1),0,#Effects)
 
  -- TODO: key instead of keyp? limit?
  -- e: effect control down
@@ -1639,7 +1060,7 @@ function KEY_CHECK()
   ETimerMode = ETimerMode + 1
  end
 
- ETimerMode = mm(ETimerMode,0,NumModes)
+ ETimerMode = clamp(ETimerMode,0,NumModes)
 
  -- u: effect divider down
  if keyp(21) == true then
@@ -1650,7 +1071,7 @@ function KEY_CHECK()
  if keyp(9) == true then
   EDivider = EDivider + 1
  end
- EDivider = mm(EDivider,-10,10)
+ EDivider = clamp(EDivider,-10,10)
 
  -- o: effect palette down
  if keyp(15) == true then
@@ -1662,7 +1083,7 @@ function KEY_CHECK()
   EPalette = EPalette + 1
  end 
 
- EPalette = mm(EPalette%(NumPalettes+1),0,NumPalettes)
+ EPalette = clamp(EPalette%(NumPalettes+1),0,NumPalettes)
  
  -- 1: effect modifier down
  if keyp(28) == true then
@@ -1676,7 +1097,7 @@ function KEY_CHECK()
   if EModifier > #Modifiers then EModifier = 1 end
  end
  
- --EModifier = mm(EModifier%(NumModifiers+1),0,NumModifiers)
+ --EModifier = clamp(EModifier%(NumModifiers+1),0,NumModifiers)
  
  -- 3: effect modifier control down
  if keyp(30) == true then
@@ -1698,7 +1119,7 @@ function KEY_CHECK()
   EMTimerMode = EMTimerMode + 1
  end
 
- EMTimerMode = mm(EMTimerMode,0,NumModes)
+ EMTimerMode = clamp(EMTimerMode,0,NumModes)
 
  -- 7: effect modifier divider down
  if keyp(34) == true then
@@ -1709,7 +1130,7 @@ function KEY_CHECK()
  if keyp(35) == true then
   EMDivider = EMDivider + 1
  end
- EMDivider = mm(EMDivider,-10,10)
+ EMDivider = clamp(EMDivider,-10,10)
 
  -- z: overlay modifier down
  if keyp(26) == true then
@@ -1723,7 +1144,7 @@ function KEY_CHECK()
   if OModifier > #Modifiers then OModifier = 1 end
  end
  
- --OModifier = mm(OModifier%(NumModifiers+1),0,NumModifiers)
+ --OModifier = clamp(OModifier%(NumModifiers+1),0,NumModifiers)
 
  -- c: overlay modifier control down
  if keyp(3) == true then
@@ -1745,7 +1166,7 @@ function KEY_CHECK()
   OMTimerMode = OMTimerMode + 1
  end
   
- OMTimerMode = mm(OMTimerMode,0,NumModes)
+ OMTimerMode = clamp(OMTimerMode,0,NumModes)
   
  -- m: overlay modifier divider down
  if keyp(13) == true then
@@ -1756,7 +1177,7 @@ function KEY_CHECK()
  if keyp(45) == true then
   OMDivider = OMDivider + 1
  end
- OMDivider = mm(OMDivider,-10,10)
+ OMDivider = clamp(OMDivider,-10,10)
   
  -- a: overlay down
  if keyp(1) == true then
@@ -1770,7 +1191,7 @@ function KEY_CHECK()
   if Overlay > #Overlays then Overlay = 1 end
  end
  
- --Overlay = mm(Overlay%(NumOverlays+1),0,NumOverlays)
+ --Overlay = clamp(Overlay%(NumOverlays+1),0,NumOverlays)
   
  -- TODO: key instead of keyp? limit?
  -- d: overlay control down
@@ -1793,7 +1214,7 @@ function KEY_CHECK()
   OTimerMode = OTimerMode + 1
  end
 
- OTimerMode = mm(OTimerMode,0,NumModes)
+ OTimerMode = clamp(OTimerMode,0,NumModes)
 
  -- j: overlay divider down
  if keyp(10) == true then
@@ -1804,7 +1225,7 @@ function KEY_CHECK()
  if keyp(11) == true then
   ODivider = ODivider + 1
  end
- ODivider = mm(ODivider,-10,10)
+ ODivider = clamp(ODivider,-10,10)
 
  -- l: overlay palette down
  if keyp(12) == true then
@@ -1816,7 +1237,7 @@ function KEY_CHECK()
   OPalette = OPalette + 1
  end 
 
- OPalette = mm(OPalette%(NumPalettes+1),0,NumPalettes)
+ OPalette = clamp(OPalette%(NumPalettes+1),0,NumPalettes)
 
  -- delete: overlay cls switch
  if keyp(52) == true then
@@ -1868,7 +1289,7 @@ function BOOT()
   FFT_BOOT()
   BEATTIME_BOOT()
   PAL_BOOT()
-  FONT_BOOT()
+  FontBoot()
 
   for _,effect in ipairs(Effects) do
     if effect.boot then
@@ -1881,41 +1302,7 @@ function BOOT()
       overlay.boot()
     end
   end
-
-  --[[
- 
- TestSheet_BOOT()
- SunBeat_BOOT()
- FujiTwist_BOOT()
- c64_BOOT()
- amiga_BOOT()
- vectrex_BOOT()
- SmokeCircles_BOOT()
- Chladni_BOOT()
- TextWarp_BOOT()
- JoyDivision_BOOT()
- Proxima_BOOT()
- Lemons_BOOT()
- RevisionBack_BOOT()
- --Bitnick_BOOT()
- --]]
 end
-
-Modifiers={
-	PIXNoise,
-	PIXZoom,
-	GridDim,
-	POSTCirc,
-	PIXMotionBlur,
-	PIXJumpBlur,
-	ROTVert,
-	ROTHorz,
-	SFTVert,
-	SFTHorz,
-	POSTSquares,
-	EvilpaulGlitch,
-	LineScratch
-}
 
 function TIC()
